Rename UnitHitsChange -> UnitDamageReceivedChange
Also rename the variables to newTotalDamage and oldTotalDamage to be
more specific in what they hold.

Add comments to help explain parts of the code.
